+++ common/db/buckets.go
@@ -11,4 +11,5 @@ const (
     BucketSyncHistory  = "sync_history"
     BucketPRIndexByID  = "pr_index_by_id"
     BucketPRIndexByRG  = "pr_index_by_rg_num"
+    BucketWebhookRetries = "webhook_retries"
 )